bash toolset: split argv detection into utils + per-command rules
Replace the single dangerous_args.py with a two-layer split:

- argv_utils.py: generic, command-agnostic helpers (getopt-style parse_argv,
  abbreviates, is_benign_redirect_target).
- command_arg_rules.py: ONLY the per-command rules (find/sort/uniq checkers +
  their constants) — the clear place to add a new command's check. Exposes just
  dangerous_argv_reason() and is_argv_checked_command(); the registry is private.

validation.py imports those functions instead of reaching into the registry.
Pure move + rename; behavior unchanged.
